Add sendActivityLog method to Rerouter class
